Separated EditProfile to another page
