+++ client/src/modules/profile/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
-// "Public" API for login 
+// "Public" API for profile
 import Profile from './components/Profile';
+import EditProfile from './components/EditProfile';
 
 
-export default { Profile }; 
+export default { EditProfile, Profile };
